+++ packages/copilot/src/copilot/container/types.ts
@@ -1,0 +1,6 @@
+export interface stepsStatusI {
+  status: string;
+  index: number;
+  component: JSX.Element;
+  title: string;
+}