+++ packages/copilot/src/copilot/steps/topUp/C14.tsx
@@ -5,22 +5,22 @@ import { StoreType } from "evmos-wallet";
 import { useSelector } from "react-redux";
 import { useTransakEvents } from "./useTransakEvents";
 
-export const TRANSAK_API_KEY = process.env.NEXT_PUBLIC_TRANSAK_API_KEY ?? "";
+export const C14_API_KEY = process.env.NEXT_PUBLIC_C14_API_KEY ?? "";
 
-export default function Transak() {
+export default function C14() {
   const wallet = useSelector((state: StoreType) => state.wallet.value);
   useTransakEvents();
 
   return (
     <div
-      data-testid="transak-widget"
+      data-testid="c14-widget"
       className="relative mx-auto mt-[25px] h-[700px] w-[300px] overflow-hidden rounded-[15px] md:h-[545px] md:w-[400px]"
     >
       <iframe
         // eslint-disable-next-line no-secrets/no-secrets
-        src={`https://global.transak.com?apiKey=${TRANSAK_API_KEY}&themeColor=ed4e33&cryptoCurrencyCode=EVMOS&network=evmos&defaultPaymentMethod=credit_debit_card&disableWalletAddressForm=true&walletAddress=${wallet.evmosAddressEthFormat}`}
-        allow="camera;microphone;fullscreen;payment"
-        style={{ height: "100%", width: "100%", border: "none" }}
+        src={`https://pay.c14.money?clientId=${C14_API_KEY}&targetAssetId=e2e0546e-b51b-4d56-9426-3aff3a2418ba&targetAssetIdLock=true&targetAddress=${wallet?.evmosAddressEthFormat}`}
+        width="100%"
+        height="100%"
       ></iframe>
     </div>
   );