+++ packages/copilot/src/locales/en.ts
@@ -5,12 +5,12 @@ export default {
     "Now that you've set up your account, you'll need Evmos tokens to access and interact with a huge catalog of decentralized apps (dApps), such as games 🎮, marketplaces 🏪, decentralized finance (DeFi) 💸, and more!",
   "topup.card.button": "Debit/Credit Card",
   "topup.crypto.button": "Cryptocurrencies",
-  "topup.intro.crypto.button.comingsoon": "Coming soon",
   "topup.onboard.title": "Onboard to Evmos",
   "topup.onboard.success.title": "Congratulations",
   "topup.onboard.success.description": "You're now ready to use your Evmos!",
   "topup.onboard.success.button.text": "Next steps",
   "topup.onboard.success.step.topup.button.text": "Continue to the dashboard",
+  "topup.onboard.card.provider": "This onboarding service is provided by",
   "setupaccount.title": "Set up your account",
   "setupaccount.description":
     "Having an account allows you to use your Evmos to interact with any decentralized applications (dApps) on Evmos.",
@@ -29,11 +29,11 @@ export default {
   "nextsteps.stakeEvmos.badge": "Advanced",
   "nextsteps.learnMore.title": "Learn more",
   /* eslint-enable no-secrets/no-secrets */
-  stakingUrl: "https://app.evmos.org/staking",
   academyFAQUrl: "https://academy.evmos.org/faq",
   "exitcopilot.title": "Exit Copilot?",
   "exitcopilot.description":
     "We advise you finish all steps if you want to get started with Evmos for the first time.",
   "exitcopilot.reject": "Exit",
   "exitcopilot.accept": "Stay",
+  "dappStore.onboard.title": "Onboard to Evmos in 5 minutes",
 } as const;