+++ packages/ui-helpers/src/launchPad/Item.tsx
@@ -1,9 +1,16 @@
 import Link from "next/link";
 import { LaunchPadItemsProps } from "./data";
+import { CLICK_ON_DAPP_INSIDE_LAUNCHER, useTracker } from "tracker";
 
 export const Item = ({ itemProps }: { itemProps: LaunchPadItemsProps }) => {
+  const { handlePreClickAction } = useTracker(CLICK_ON_DAPP_INSIDE_LAUNCHER);
+
+  const handleOnClick = () => {
+    handlePreClickAction({ dApp: itemProps.text });
+  };
   return (
     <Link
+      onClick={handleOnClick}
       href={itemProps.href}
       rel="noopener noreferrer"
       className="text-pearl flex flex-col items-center"